chore(core): refactor agent.execute fn workflow
