+++ .changeset/little-streets-stick.md
@@ -1,0 +1,5 @@
+---
+'@mastra/core': patch
+---
+
+Refactor agent.#execute fn workflow to make code easier to follow.
