+++ packages/core/src/agent/agent.ts
@@ -3,10 +3,9 @@ import type { WritableStream } from 'stream/web';
 import type { CoreMessage, StreamObjectResult, TextPart, Tool, UIMessage } from 'ai';
 import deepEqual from 'fast-deep-equal';
 import type { JSONSchema7 } from 'json-schema';
-import { z } from 'zod';
-import type { ZodSchema } from 'zod';
+import type { z, ZodSchema } from 'zod';
 import type { MastraPrimitives, MastraUnion } from '../action';
-import { AISpanType, getOrCreateSpan, getValidTraceId, InternalSpans } from '../ai-tracing';
+import { AISpanType, getOrCreateSpan, getValidTraceId } from '../ai-tracing';
 import type { AISpan, TracingContext, TracingOptions, TracingProperties } from '../ai-tracing';
 import { MastraBase } from '../base';
 import { MastraError, ErrorDomain, ErrorCategory } from '../error';
@@ -28,7 +27,6 @@ import type {
   StreamTextResult,
 } from '../llm/model/base.types';
 import { MastraLLMVNext } from '../llm/model/model.loop';
-import type { ModelLoopStreamArgs } from '../llm/model/model.loop.types';
 import type { TripwireProperties, MastraLanguageModel, MastraLanguageModelV2 } from '../llm/model/shared.types';
 import { RegisteredLogger } from '../logger';
 import { networkLoop } from '../loop/network';
@@ -50,7 +48,6 @@ import { runScorer } from '../scores/hooks';
 import type { AISDKV5OutputStream } from '../stream';
 import type { MastraModelOutput } from '../stream/base/output';
 import type { OutputSchema } from '../stream/base/schema';
-import { ChunkFrom } from '../stream/types';
 import type { ChunkType } from '../stream/types';
 import { InstrumentClass } from '../telemetry';
 import { Telemetry } from '../telemetry/telemetry';
@@ -61,7 +58,6 @@ import { makeCoreTool, createMastraProxy, ensureToolProperties } from '../utils'
 import type { ToolOptions } from '../utils';
 import type { CompositeVoice } from '../voice';
 import { DefaultVoice } from '../voice';
-import { createStep, createWorkflow } from '../workflows';
 import type { Workflow } from '../workflows';
 import { agentToStep, LegacyStep as Step } from '../workflows/legacy';
 import type { AgentExecutionOptions, InnerAgentExecutionOptions, MultiPrimitiveExecutionOptions } from './agent.types';
@@ -3014,515 +3010,47 @@ export class Agent<
       this.logger.debug(`[Agents:${this.name}] - Starting generation`, { runId });
     }
 
-    const prepareToolsStep = createStep({
-      id: 'prepare-tools-step',
-      inputSchema: z.any(),
-      outputSchema: z.object({
-        convertedTools: z.record(z.string(), z.any()),
-      }),
-      execute: async () => {
-        const toolEnhancements = [
-          // toolsets
-          options?.toolsets && Object.keys(options?.toolsets || {}).length > 0
-            ? `toolsets present (${Object.keys(options?.toolsets || {}).length} tools)`
-            : undefined,
+    // Import the workflow creator function
+    const { createPrepareStreamWorkflow } = await import('./workflows/prepare-stream');
 
-          // memory tools
-          memory && resourceId ? 'memory and resourceId available' : undefined,
-        ]
-          .filter(Boolean)
-          .join(', ');
+    // Create a capabilities object with bound methods
+    const capabilities = {
+      agentName: this.name,
+      agentInstructions: instructions,
+      logger: this.logger,
+      getMemory: this.getMemory.bind(this),
+      getModel: this.getModel.bind(this),
+      generateMessageId: this.#mastra?.generateId?.bind(this.#mastra) || (() => randomUUID()),
+      _agentNetworkAppend:
+        '_agentNetworkAppend' in this
+          ? Boolean((this as unknown as { _agentNetworkAppend: unknown })._agentNetworkAppend)
+          : undefined,
+      saveStepMessages: this.saveStepMessages.bind(this),
+      convertTools: this.convertTools.bind(this),
+      getMemoryMessages: this.getMemoryMessages.bind(this),
+      runInputProcessors: this.__runInputProcessors.bind(this),
+      executeOnFinish: this.#executeOnFinish.bind(this),
+      outputProcessors: this.#outputProcessors,
+      llm,
+    };
 
-        this.logger.debug(`[Agent:${this.name}] - Enhancing tools: ${toolEnhancements}`, {
-          runId,
-          toolsets: options?.toolsets ? Object.keys(options?.toolsets) : undefined,
-          clientTools: options?.clientTools ? Object.keys(options?.clientTools) : undefined,
-          hasMemory: !!memory,
-          hasResourceId: !!resourceId,
-        });
-
-        const threadId = threadFromArgs?.id;
-
-        const convertedTools = await this.convertTools({
-          toolsets: options?.toolsets,
-          clientTools: options?.clientTools,
-          threadId,
-          resourceId,
-          runId,
-          runtimeContext,
-          tracingContext: { currentSpan: agentAISpan },
-          writableStream: options.writableStream,
-          methodType,
-          format,
-        });
-
-        return {
-          convertedTools,
-        };
-      },
+    // Create the workflow with all necessary context
+    const executionWorkflow = createPrepareStreamWorkflow({
+      capabilities,
+      options: { ...options, methodType },
+      threadFromArgs,
+      resourceId,
+      runId,
+      runtimeContext,
+      agentAISpan: agentAISpan!,
+      methodType,
+      format: format as FORMAT,
+      instructions,
+      memoryConfig,
+      memory,
+      saveQueueManager,
+      returnScorerData: options.returnScorerData,
     });
-
-    const prepareMemory = createStep({
-      id: 'prepare-memory-step',
-      inputSchema: z.any(),
-      outputSchema: z.object({
-        threadExists: z.boolean(),
-        thread: z.any(),
-        messageList: z.any(),
-        tripwire: z.boolean().optional(),
-        tripwireReason: z.string().optional(),
-      }),
-      execute: async ({ tracingContext }) => {
-        const thread = threadFromArgs;
-        const messageList = new MessageList({
-          threadId: thread?.id,
-          resourceId,
-          generateMessageId: this.#mastra?.generateId?.bind(this.#mastra),
-          // @ts-ignore Flag for agent network messages
-          _agentNetworkAppend: this._agentNetworkAppend,
-        })
-          .addSystem({
-            role: 'system',
-            content: instructions || `${this.instructions}.`,
-          })
-          .add(options.context || [], 'context');
-
-        if (!memory || (!thread?.id && !resourceId)) {
-          messageList.add(options.messages, 'user');
-          const { tripwireTriggered, tripwireReason } = await this.__runInputProcessors({
-            runtimeContext,
-            tracingContext,
-            messageList,
-          });
-          return {
-            threadExists: false,
-            thread: undefined,
-            messageList,
-            ...(tripwireTriggered && {
-              tripwire: true,
-              tripwireReason,
-            }),
-          };
-        }
-        if (!thread?.id || !resourceId) {
-          const mastraError = new MastraError({
-            id: 'AGENT_MEMORY_MISSING_RESOURCE_ID',
-            domain: ErrorDomain.AGENT,
-            category: ErrorCategory.USER,
-            details: {
-              agentName: this.name,
-              threadId: thread?.id || '',
-              resourceId: resourceId || '',
-            },
-            text: `A resourceId and a threadId must be provided when using Memory. Saw threadId "${thread?.id}" and resourceId "${resourceId}"`,
-          });
-          this.logger.trackException(mastraError);
-          this.logger.error(mastraError.toString());
-          throw mastraError;
-        }
-        const store = memory.constructor.name;
-        this.logger.debug(
-          `[Agent:${this.name}] - Memory persistence enabled: store=${store}, resourceId=${resourceId}`,
-          {
-            runId,
-            resourceId,
-            threadId: thread?.id,
-            memoryStore: store,
-          },
-        );
-
-        let threadObject: StorageThreadType | undefined = undefined;
-        const existingThread = await memory.getThreadById({ threadId: thread?.id });
-
-        if (existingThread) {
-          if (
-            (!existingThread.metadata && thread.metadata) ||
-            (thread.metadata && !deepEqual(existingThread.metadata, thread.metadata))
-          ) {
-            threadObject = await memory.saveThread({
-              thread: { ...existingThread, metadata: thread.metadata },
-              memoryConfig,
-            });
-          } else {
-            threadObject = existingThread;
-          }
-        } else {
-          threadObject = await memory.createThread({
-            threadId: thread?.id,
-            metadata: thread.metadata,
-            title: thread.title,
-            memoryConfig,
-            resourceId,
-            saveThread: false,
-          });
-        }
-
-        const config = memory.getMergedThreadConfig(memoryConfig || {});
-        const hasResourceScopeSemanticRecall =
-          typeof config?.semanticRecall === 'object' && config?.semanticRecall?.scope === 'resource';
-        let [memoryMessages, memorySystemMessage] = await Promise.all([
-          existingThread || hasResourceScopeSemanticRecall
-            ? this.getMemoryMessages({
-                resourceId,
-                threadId: threadObject.id,
-                vectorMessageSearch: new MessageList().add(options.messages, `user`).getLatestUserContent() || '',
-                memoryConfig,
-                runtimeContext,
-              })
-            : [],
-          memory.getSystemMessage({ threadId: threadObject.id, resourceId, memoryConfig }),
-        ]);
-
-        this.logger.debug('Fetched messages from memory', {
-          threadId: threadObject.id,
-          runId,
-          fetchedCount: memoryMessages.length,
-        });
-
-        // So the agent doesn't get confused and start replying directly to messages
-        // that were added via semanticRecall from a different conversation,
-        // we need to pull those out and add to the system message.
-        const resultsFromOtherThreads = memoryMessages.filter(m => m.threadId !== threadObject.id);
-        if (resultsFromOtherThreads.length && !memorySystemMessage) {
-          memorySystemMessage = ``;
-        }
-        if (resultsFromOtherThreads.length) {
-          memorySystemMessage += `\nThe following messages were remembered from a different conversation:\n<remembered_from_other_conversation>\n${(() => {
-            let result = ``;
-
-            const messages = new MessageList().add(resultsFromOtherThreads, 'memory').get.all.v1();
-            let lastYmd: string | null = null;
-            for (const msg of messages) {
-              const date = msg.createdAt;
-              const year = date.getUTCFullYear();
-              const month = date.toLocaleString('default', { month: 'short' });
-              const day = date.getUTCDate();
-              const ymd = `${year}, ${month}, ${day}`;
-              const utcHour = date.getUTCHours();
-              const utcMinute = date.getUTCMinutes();
-              const hour12 = utcHour % 12 || 12;
-              const ampm = utcHour < 12 ? 'AM' : 'PM';
-              const timeofday = `${hour12}:${utcMinute < 10 ? '0' : ''}${utcMinute} ${ampm}`;
-
-              if (!lastYmd || lastYmd !== ymd) {
-                result += `\nthe following messages are from ${ymd}\n`;
-              }
-              result += `Message ${msg.threadId && msg.threadId !== threadObject.id ? 'from previous conversation' : ''} at ${timeofday}: ${JSON.stringify(msg)}`;
-
-              lastYmd = ymd;
-            }
-            return result;
-          })()}\n<end_remembered_from_other_conversation>`;
-        }
-
-        if (memorySystemMessage) {
-          messageList.addSystem(memorySystemMessage, 'memory');
-        }
-
-        messageList
-          .add(
-            memoryMessages.filter(m => m.threadId === threadObject.id), // filter out messages from other threads. those are added to system message above
-            'memory',
-          )
-          // add new user messages to the list AFTER remembered messages to make ordering more reliable
-          .add(options.messages, 'user');
-
-        const { tripwireTriggered, tripwireReason } = await this.__runInputProcessors({
-          runtimeContext,
-          tracingContext,
-          messageList,
-        });
-
-        const systemMessages = messageList.getSystemMessages();
-
-        const systemMessage =
-          [...systemMessages, ...messageList.getSystemMessages('memory')]?.map(m => m.content)?.join(`\n`) ?? undefined;
-
-        const processedMemoryMessages = await memory.processMessages({
-          // these will be processed
-          messages: messageList.get.remembered.v1() as CoreMessage[],
-          // these are here for inspecting but shouldn't be returned by the processor
-          // - ex TokenLimiter needs to measure all tokens even though it's only processing remembered messages
-          newMessages: messageList.get.input.v1() as CoreMessage[],
-          systemMessage,
-          memorySystemMessage: memorySystemMessage || undefined,
-        });
-
-        const processedList = new MessageList({
-          threadId: threadObject.id,
-          resourceId,
-          generateMessageId: this.#mastra?.generateId?.bind(this.#mastra),
-          // @ts-ignore Flag for agent network messages
-          _agentNetworkAppend: this._agentNetworkAppend,
-        })
-          .addSystem(instructions || `${this.instructions}.`)
-          .addSystem(memorySystemMessage)
-          .addSystem(systemMessages)
-          .add(options.context || [], 'context')
-          .add(processedMemoryMessages, 'memory')
-          .add(messageList.get.input.v2(), 'user');
-
-        return {
-          thread: threadObject,
-          messageList: processedList,
-          // add old processed messages + new input messages
-          ...(tripwireTriggered && {
-            tripwire: true,
-            tripwireReason,
-          }),
-          threadExists: !!existingThread,
-        };
-      },
-    });
-
-    const streamStep = createStep({
-      id: 'stream-text-step',
-      inputSchema: z.any(),
-      outputSchema: z.any(),
-      execute: async ({ inputData, tracingContext }) => {
-        this.logger.debug(`Starting agent ${this.name} llm stream call`, {
-          runId,
-        });
-
-        const outputProcessors =
-          inputData.outputProcessors ||
-          (this.#outputProcessors
-            ? typeof this.#outputProcessors === 'function'
-              ? await this.#outputProcessors({
-                  runtimeContext: inputData.runtimeContext || new RuntimeContext(),
-                })
-              : this.#outputProcessors
-            : []);
-
-        const streamResult = llm.stream({
-          ...inputData,
-          outputProcessors,
-          returnScorerData: options.returnScorerData,
-          tracingContext,
-          _internal: {
-            generateId: inputData.experimental_generateMessageId || this.#mastra?.generateId?.bind(this.#mastra),
-          },
-        });
-
-        if (format === 'aisdk') {
-          return streamResult.aisdk.v5;
-        }
-
-        return streamResult;
-      },
-    });
-
-    const executionWorkflow = createWorkflow({
-      id: 'execution-workflow',
-      inputSchema: z.any(),
-      outputSchema: z.any(),
-      steps: [prepareToolsStep, prepareMemory],
-      options: {
-        tracingPolicy: {
-          // mark all workflow spans related to the
-          // VNext execution as internal
-          internal: InternalSpans.WORKFLOW,
-        },
-      },
-    })
-      .parallel([prepareToolsStep, prepareMemory])
-      .map(async ({ inputData, bail }) => {
-        const result = {
-          ...options,
-          tools: inputData['prepare-tools-step'].convertedTools as Record<string, Tool>,
-          runId,
-          temperature: options.modelSettings?.temperature,
-          toolChoice: options.toolChoice,
-          thread: inputData['prepare-memory-step'].thread,
-          threadId: inputData['prepare-memory-step'].thread?.id,
-          resourceId,
-          runtimeContext,
-          onStepFinish: async (props: any) => {
-            if (options.savePerStep) {
-              if (!inputData['prepare-memory-step'].threadExists && memory && inputData['prepare-memory-step'].thread) {
-                await memory.createThread({
-                  threadId: inputData['prepare-memory-step'].thread?.id,
-                  title: inputData['prepare-memory-step'].thread?.title,
-                  metadata: inputData['prepare-memory-step'].thread?.metadata,
-                  resourceId: inputData['prepare-memory-step'].thread?.resourceId,
-                  memoryConfig,
-                });
-
-                inputData['prepare-memory-step'].threadExists = true;
-              }
-
-              await this.saveStepMessages({
-                saveQueueManager,
-                result: props,
-                messageList: inputData['prepare-memory-step'].messageList,
-                threadId: inputData['prepare-memory-step'].thread?.id,
-                memoryConfig,
-                runId,
-              });
-            }
-
-            return options.onStepFinish?.({ ...props, runId });
-          },
-          ...(inputData['prepare-memory-step'].tripwire && {
-            tripwire: inputData['prepare-memory-step'].tripwire,
-            tripwireReason: inputData['prepare-memory-step'].tripwireReason,
-          }),
-        } as any;
-
-        // Check for tripwire and return early if triggered
-        if (result.tripwire) {
-          // Return a promise that resolves immediately with empty result
-          const emptyResult = {
-            textStream: (async function* () {
-              // Empty async generator - yields nothing
-            })(),
-            fullStream: new globalThis.ReadableStream({
-              start(controller: any) {
-                controller.enqueue({
-                  type: 'tripwire',
-                  runId: result.runId,
-                  from: ChunkFrom.AGENT,
-                  payload: {
-                    tripwireReason: result.tripwireReason,
-                  },
-                });
-                controller.close();
-              },
-            }),
-            objectStream: new globalThis.ReadableStream({
-              start(controller: any) {
-                controller.close();
-              },
-            }),
-            text: Promise.resolve(''),
-            usage: Promise.resolve({ inputTokens: 0, outputTokens: 0, totalTokens: 0 }),
-            finishReason: Promise.resolve('other'),
-            tripwire: true,
-            tripwireReason: result.tripwireReason,
-            response: {
-              id: randomUUID(),
-              timestamp: new Date(),
-              modelId: 'tripwire',
-              messages: [],
-            },
-            toolCalls: Promise.resolve([]),
-            toolResults: Promise.resolve([]),
-            warnings: Promise.resolve(undefined),
-            request: {
-              body: JSON.stringify({ messages: [] }),
-            },
-            object: undefined,
-            experimental_output: undefined,
-            steps: undefined,
-            experimental_providerMetadata: undefined,
-          };
-
-          return bail(emptyResult);
-        }
-
-        let effectiveOutputProcessors =
-          options.outputProcessors ||
-          (this.#outputProcessors
-            ? typeof this.#outputProcessors === 'function'
-              ? await this.#outputProcessors({
-                  runtimeContext: result.runtimeContext!,
-                })
-              : this.#outputProcessors
-            : []);
-
-        // Handle structuredOutput option by creating an StructuredOutputProcessor
-        if (options.structuredOutput) {
-          const agentModel = await this.getModel({ runtimeContext: result.runtimeContext! });
-          const structuredProcessor = new StructuredOutputProcessor(options.structuredOutput, agentModel);
-          effectiveOutputProcessors = effectiveOutputProcessors
-            ? [...effectiveOutputProcessors, structuredProcessor]
-            : [structuredProcessor];
-        }
-
-        const loopOptions: ModelLoopStreamArgs<any, OUTPUT> = {
-          runtimeContext: result.runtimeContext!,
-          tracingContext: { currentSpan: agentAISpan },
-          runId,
-          toolChoice: result.toolChoice,
-          tools: result.tools,
-          resourceId: result.resourceId,
-          threadId: result.threadId,
-          structuredOutput: result.structuredOutput,
-          stopWhen: result.stopWhen,
-          maxSteps: result.maxSteps,
-          providerOptions: result.providerOptions,
-          options: {
-            ...(options.prepareStep && { prepareStep: options.prepareStep }),
-            onFinish: async (payload: any) => {
-              if (payload.finishReason === 'error') {
-                this.logger.error('Error in agent stream', {
-                  error: payload.error,
-                  runId,
-                });
-                return;
-              }
-
-              const messageList = inputData['prepare-memory-step'].messageList as MessageList;
-
-              try {
-                const outputText = messageList.get.all
-                  .core()
-                  .map(m => m.content)
-                  .join('\n');
-
-                await this.#executeOnFinish({
-                  result: payload,
-                  outputText,
-                  instructions,
-                  thread: result.thread,
-                  threadId: result.threadId,
-                  readOnlyMemory: options.memory?.readOnly,
-                  resourceId,
-                  memoryConfig,
-                  runtimeContext,
-                  agentAISpan: agentAISpan,
-                  runId,
-                  messageList,
-                  threadExists: inputData['prepare-memory-step'].threadExists,
-                  structuredOutput: !!options.output,
-                  saveQueueManager,
-                  overrideScorers: options.scorers,
-                });
-              } catch (e) {
-                this.logger.error('Error saving memory on finish', {
-                  error: e,
-                  runId,
-                });
-              }
-
-              await options?.onFinish?.({
-                ...payload,
-                runId,
-                messages: messageList.get.response.aiV5.model(),
-                usage: payload.usage,
-                totalUsage: payload.totalUsage,
-              });
-            },
-            onStepFinish: result.onStepFinish,
-            onChunk: options.onChunk,
-            onError: options.onError,
-            onAbort: options.onAbort,
-            activeTools: options.activeTools,
-            abortSignal: options.abortSignal,
-          },
-          output: options.output,
-          outputProcessors: effectiveOutputProcessors,
-          modelSettings: {
-            temperature: 0,
-            ...(options.modelSettings || {}),
-          },
-          messageList: inputData['prepare-memory-step'].messageList,
-        };
-
-        return loopOptions;
-      })
-      .then(streamStep)
-      .commit();
 
     const run = await executionWorkflow.createRunAsync();
     const result = await run.start({ tracingContext: { currentSpan: agentAISpan } });
